Add option to scp HTML copy of console log.
Add functionality to the SCP plugin that allows it to copy the Jenkins
Console log (as HTML) to a remote host. Also add flag to copy SCP source
even when the build has failed (useful when you want to copy the console
log on for all builds).

This copied HTML console log will include timestamps if you are logging
them, and will list the build result (up to the point of when the copy
is made) at the end of the log.