fix uploading of build console logs to remote server
Fixes issue JENKINS-21138

An exception occurs during console log upload causing only partial logs
to be copied to the remote server.  The problem is that the session object
is not thread safe.  To fix this problem we need to move the session
object into the ConsoleRunnable thread.  Now the console upload always
runs in a completely detached thread.

Notes:
1. Needed to update jenkins version to 1.430 because 1.403 would not build
   anymore.  I think 1.430 should be conservative enought to be compatible
   with most of the existing jenkins installs.
2. Updated com.jcraft.jsch depenency to version 1.50